Render stats use u64.
CommandList set to be discard if the uniform buffer is resized.
DescriptorAllocator now hashes the upload data to be used as a cache. If the data has been uploaded before reuse it. Reduces memory upload amount and usage. Adds some CPU cost. On Sponza with curtains default camera placement, upload data amount went from 500 KB to around 157 KB. CPU cost was very very small (will need to keep an eye on this).